predict ghost zlevel moving
