+++ Content.Client/_CE/ZLevels/Ghost/CEClientZLevelGhostMoverSystem.cs
@@ -1,0 +1,12 @@
+/*
+ * This file is sublicensed under MIT License
+ * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
+ */
+
+using Content.Shared._CE.ZLevels.Ghost;
+
+namespace Content.Client._CE.ZLevels.Ghost;
+
+public sealed class CEClientZLevelGhostMoverSystem : CESharedZLevelGhostMoverSystem
+{
+}
