+++ Content.Server/_CE/ZLevels/Ghost/CEZLevelGhostMoverSystem.cs
@@ -3,17 +3,14 @@
  * https://github.com/space-wizards/space-station-14/blob/master/LICENSE.TXT
  */
 
-using Content.Shared._CE.ZLevels.Core.Components;
-using Content.Shared._CE.ZLevels.Core.EntitySystems;
 using Content.Shared._CE.ZLevels.Ghost;
 using Content.Shared.Actions;
 
 namespace Content.Server._CE.ZLevels.Ghost;
 
-public sealed class CEZLevelGhostMoverSystem : EntitySystem
+public sealed class CEZLevelGhostMoverSystem : CESharedZLevelGhostMoverSystem
 {
     [Dependency] private readonly SharedActionsSystem _actions = default!;
-    [Dependency] private readonly CESharedZLevelsSystem _zLevel = default!;
 
     public override void Initialize()
     {
@@ -21,8 +18,6 @@ public sealed class CEZLevelGhostMoverSystem : EntitySystem
 
         SubscribeLocalEvent<CEZLevelGhostMoverComponent, MapInitEvent>(OnMapInit);
         SubscribeLocalEvent<CEZLevelGhostMoverComponent, ComponentRemove>(OnRemove);
-        SubscribeLocalEvent<CEZLevelGhostMoverComponent, CEZLevelActionUp>(OnZLevelUp);
-        SubscribeLocalEvent<CEZLevelGhostMoverComponent, CEZLevelActionDown>(OnZLevelDown);
     }
 
     private void OnMapInit(Entity<CEZLevelGhostMoverComponent> ent, ref MapInitEvent args)
@@ -35,21 +30,5 @@ public sealed class CEZLevelGhostMoverSystem : EntitySystem
     {
         _actions.RemoveAction(ent.Comp.ZLevelUpActionEntity);
         _actions.RemoveAction(ent.Comp.ZLevelDownActionEntity);
-    }
-
-    private void OnZLevelDown(Entity<CEZLevelGhostMoverComponent> ent, ref CEZLevelActionDown args)
-    {
-        if (args.Handled)
-            return;
-
-        args.Handled = _zLevel.TryMoveDown(ent);
-    }
-
-    private void OnZLevelUp(Entity<CEZLevelGhostMoverComponent> ent, ref CEZLevelActionUp args)
-    {
-        if (args.Handled)
-            return;
-
-        args.Handled = _zLevel.TryMoveUp(ent);
     }
 }
